Add run_id_str option: `-i <run_id_str>`.
This can be used to replace a field in the command's ARGV with the
run ID whenever the option string matches.

For example,

    $ autoclave -i % -s examples/fail_if_argv1_eq_5 %

Will run examples/fail_if_argv1_eq_5 as

    examples/fail_if_argv1_eq_5 0
    examples/fail_if_argv1_eq_5 1
    examples/fail_if_argv1_eq_5 2
    examples/fail_if_argv1_eq_5 3
    examples/fail_if_argv1_eq_5 4
    examples/fail_if_argv1_eq_5 5 (which fails)

This could be used to base a RNG seed on the run ID, save other
output artifacts from the program in a distinct path, etc.