Change '-w wait_msec' to '-m min_runtime_msec'.
Instead of blocking for WAIT_MSEC after each run, only wait when
necessary to pad the run time to MIN_RUNTIME_MSEC. The delay default was
there so a very short-lived process won't unexpectedly spin in a tight
loop, but by default it shouldn't slow down processes that take a
perceptible amount of time.

Also reduce the default to 50 msec (from 100 msec), and update
the documentation.

Thanks to Josh Triplett for the feature suggestion.

Closes #8.