Wrap use of SIGRTMAX in an `#ifdef` check -- not defined on Darwin.
Darwin / macOS doesn't support realtime signals, this constant isn't
defined (despite coming from POSIX.1).

Fall back on a hardcoded count that should be a reasonable default.
If the signal ID is invalid, kill will reject it with EINVAL, which
is handled properly. Reporting an invalid signal upfront would be
better, but it seems difficult to do portably.

Fixes #12.